Moving of EventSourceFactory to its own module
- And some additional fixes for tests
