+++ ctapipe/io/tests/test_event_source.py
@@ -1,9 +1,5 @@
-from os.path import join, dirname
 from ctapipe.utils import get_dataset
-from ctapipe.io.hessioeventsource import HESSIOEventSource
-from ctapipe.io.eventsource import EventSource, EventSourceFactory
-import pytest
-from traitlets import TraitError
+from ctapipe.io.eventsource import EventSource
 
 
 def test_construct():
@@ -14,6 +10,7 @@ def test_construct():
     raise TypeError("EventSource should raise a TypeError when "
                     "instantiated due to its abstract methods")
 
+
 class DummyReader(EventSource):
     """
     Simple working EventSource
@@ -21,7 +18,8 @@ class DummyReader(EventSource):
     def _generator(self):
         return range(len(self.input_url))
 
-    def is_compatible(self, path):
+    @staticmethod
+    def is_compatible(file_path):
         return False
 
 
@@ -35,81 +33,3 @@ def test_is_iterable():
     test_reader = DummyReader(None, None, input_url=dataset)
     for _ in test_reader:
         pass
-
-
-def test_factory_subclasses():
-    factory= EventSourceFactory(None,None)
-    assert len(factory.subclass_names)>0
-
-def test_factory():
-    dataset = get_dataset("gamma_test.simtel.gz")
-    reader = EventSourceFactory.produce(None, None, input_url=dataset)
-    assert reader.__class__.__name__ == "HESSIOEventSource"
-    assert reader.input_url == dataset
-
-
-def test_factory_different_file():
-    dataset = get_dataset("gamma_test_large.simtel.gz")
-    reader = EventSourceFactory.produce(None, None, input_url=dataset)
-    assert reader.__class__.__name__ == "HESSIOEventSource"
-    assert reader.input_url == dataset
-
-
-def test_factory_from_reader():
-    dataset = get_dataset("gamma_test.simtel.gz")
-    reader = EventSourceFactory.produce(
-        None, None,
-        reader='HESSIOEventSource',
-        input_url=dataset
-    )
-    assert reader.__class__.__name__ == "HESSIOEventSource"
-    assert reader.input_url == dataset
-
-
-def test_factory_unknown_file_format():
-    with pytest.raises(ValueError):
-        dataset = get_dataset("optics.ecsv.txt")
-        reader = EventSourceFactory.produce(None, None, input_url=dataset)
-
-
-def test_factory_unknown_reader():
-    with pytest.raises(TraitError):
-        dataset = get_dataset("gamma_test.simtel.gz")
-        reader = EventSourceFactory.produce(
-            None, None,
-            reader='UnknownFileReader',
-            input_url=dataset
-        )
-
-
-def test_factory_incompatible_file():
-    dataset = get_dataset("optics.ecsv.txt")
-    reader = EventSourceFactory.produce(
-        None, None,
-        reader='HESSIOEventSource',
-        input_url=dataset
-    )
-    event_list = [event for event in reader]
-    assert len(event_list) == 0
-    # TODO: Need better test for this, why does pyhessio not throw an error?
-
-
-def test_factory_nonexistant_file():
-    with pytest.raises(FileNotFoundError):
-        dataset = "/fake_path/fake_file.fake_extension"
-        reader = EventSourceFactory.produce(
-            None, None,
-            reader='HESSIOEventSource',
-            input_url=dataset
-        )
-
-
-def test_factory_incorrect_use():
-    with pytest.raises(AssertionError):
-        dataset = get_dataset("gamma_test_large.simtel.gz")
-        factory = EventSourceFactory(
-            None, None,
-            input_url=dataset
-        )
-        reader = factory.produce(None, None)
-        assert reader.input_url == dataset
